Added Method to get workschedule and moved DTOs and Service Implementation to correct folders
