+++ Backend/Schichtpilot/Models/DTOs/QueryableSchedules.cs
@@ -1,0 +1,7 @@
+namespace Schichtpilot.Models.DTOs;
+
+public class QueryableSchedules
+{
+    public int Count { get; set; }
+    public IEnumerable<WorkScheduleShortDto> WorkSchedules { get; set; }
+}
