+++ Backend/Schichtpilot/Models/Enums/ScheduleStatusEnum.cs
@@ -1,0 +1,10 @@
+namespace Schichtpilot.Models.Enums;
+
+public enum ScheduleStatusEnum
+{
+    All,
+    Active,
+    Inactive,
+    Valid,
+    Invalid
+}
